requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.119.0
google-api-python-client==2.151.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
h11==0.16.0
idna==3.11
inflection==0.5.1
pyairtable==3.2.0
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
reportlab==3.6.13
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
python-docx==1.1.2
